// src/shared/enums/datasheet-enums/multi-select-actions.ts
export enum MultiSelectActionTypes {
  TOGGLE_DROPDOWN = 'TOGGLE_DROPDOWN',
  SET_SEARCH_TERM = 'SET_SEARCH_TERM',
  ADD_SELECTED_OPTION = 'ADD_SELECTED_OPTION',
  REMOVE_SELECTED_OPTION = 'REMOVE_SELECTED_OPTION',
  SET_CURRENT_PHRASE = 'SET_CURRENT_PHRASE',
  SET_CURRENT_PRODUCT = 'SET_CURRENT_PRODUCT',
  INITIALIZE_DATA = 'INITIALIZE_DATA',
  RESET_PHRASE = 'RESET_PHRASE',
  PROCESS_NUMERIC_PHRASE = 'PROCESS_NUMERIC_PHRASE',
}